requirements.txt
streamlit
pytesseract
pdf2image
pillow
pymupdf
azure-storage-blob
pyyaml
requests
pandas
python-dateutil